+++ test/libsolidity/syntaxTests/immutable/function_initialization_reading.sol
@@ -1,0 +1,7 @@
+contract C {
+    uint immutable x = f();
+
+    function f() public pure returns (uint) { return 3 + x; }
+}
+// ----
+// TypeError: (99-100): Immutable variables cannot be read in the constructor or any function or modifier called by it.