+++ test/libsolidity/syntaxTests/immutable/getter.sol
@@ -1,5 +1,5 @@
 contract C {
-    uint immutable public x;
+    uint immutable public x = 1;
 }
 // ----
 // UnimplementedFeatureError: NONE