+++ test/libsolidity/syntaxTests/immutable/reading_during_statevar_init.sol
@@ -1,0 +1,6 @@
+contract C {
+    uint immutable x = 0;
+    uint y = x;
+}
+// ----
+// TypeError: (52-53): Immutable variables cannot be read in the constructor or any function or modifier called by it.