+++ test/libsolidity/syntaxTests/immutable/uninitialized.sol
@@ -1,0 +1,5 @@
+contract C {
+    uint immutable x;
+}
+// ----
+// TypeError: (0-36): Construction controlflow ends without initializing all immutable state variables.